Fix protocol bug causing user status to show up incorrectly (#202)
At least for the Japanese language client "Project 64k 0.13 (01 Aug 2003)", some statuses are switched, so that when you join the server it won't show the player statuses correctly.

To enable switching the statuses for this client, set server.switchStatusBytesForBuggyClient=true (default false).

This also puts the "Ping timeout" to lowercase which triggers the ping timeout text to show up in pink. lol.